requirements.txt
fastapi
uvicorn[standard]
python-pptx
reportlab
openai
chromadb
